2018 DCWS sample indicators

This is a subset of the 2018 DataHaven Community Wellbeing Survey data for Greater New Haven covering several commonly used indicators, split into separate data frames for each topic but with the same format.

Usage

chronic_disease

community_cohesion

fin_insecurity

walkability

Format

Data frames with 4 columns:

question

Abbreviated indicator

category

Category: gender, age, etc.

group

Group: male, female, ages 18–34, etc.

value

Percentage value

An object of class tbl_df (inherits from tbl, data.frame) with 69 rows and 4 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 69 rows and 4 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 69 rows and 4 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 46 rows and 4 columns.

Source

2018 DataHaven Community Wellbeing Survey

Details

Detail and meaning of values for each dataset and question:

Dataset	Question	Definition / response

chronic_disease	hypertension	Received hypertension diagnosis

chronic_disease	asthma	Received asthma diagnosis

chronic_disease	diabetes	Received diabetes diagnosis

community_cohesion	positive_role_models	Strongly / somewhat agree local youth have positive role models

community_cohesion	satisfied_with_area	Satisfied with area

community_cohesion	trust_neighbors	Strongly / somewhat agree neighbors can be trusted

fin_insecurity	food_insecurity	Unable to afford food in past 12 months

fin_insecurity	housing_insecurity	Unable to afford adequate housing in past 12 months

fin_insecurity	no_bank_account	Didn't have bank account in past 12 months

walkability	safe_biking	Strongly / somewhat agree there are safe places to bike nearby

walkability	safe_sidewalks	Strongly / somewhat agree there are safe sidewalks & crosswalks nearby