2015-2024 DCWS sample trends

This is a subset of questions from both the 2015 & 2024 DataHaven Community Wellbeing Surveys for Greater New Haven.

Usage

cws_trend

Format

A data frame with 36 rows and 5 variables:

question

Indicator: local government is responsive, police approval, parks in good condition

year

Year, either 2015 or 2024

category

Category: gender, age, etc.

group

Group: male, female, ages 18-34, etc.

value

Percentage value

Source

2015 & 2024 DataHaven Community Wellbeing Survey

Examples

head(cws_trend)
#> # A tibble: 6 × 5
#>   question               year category group             value
#>   <fct>                 <dbl> <fct>    <fct>             <dbl>
#> 1 local_govt_responsive  2015 Total    Connecticut        0.49
#> 2 local_govt_responsive  2015 Total    Greater New Haven  0.47
#> 3 local_govt_responsive  2015 Age      Ages 18-34         0.42
#> 4 local_govt_responsive  2015 Age      Ages 35-49         0.4 
#> 5 local_govt_responsive  2015 Age      Ages 50-64         0.52
#> 6 local_govt_responsive  2015 Age      Ages 65+           0.59