Parameterized, project-based file paths

Quick wrapper around stringr::str_glue and here::here to fill a small recurring need: filling in file paths based on some parameter, but building them relative to the project root.

Usage

glue_here(x, ..., .snake = TRUE)

Arguments

x

Character vector of paths, relative to the project root, with glue-style encodings.

...

Additional named arguments to fill into glue brackets

.snake

Boolean, whether text should also be converted to snakecase. Default: TRUE

Value

A character vector of absolute paths with glue strings filled in.

See also

Examples

year <- 2025
loc <- "New Haven"
glue_here(c("input_data/demo/{loc}_table_{year}.csv",
            "{year}_headings.txt"))
#> Error in purrr::map(x, function(x0) {    glue_here_once_(x0, ..., .snake = .snake, .envir = .GlobalEnv)}): ℹ In index: 1.
#> Caused by error:
#> ! Failed to evaluate glue component {loc}
#> Caused by error:
#> ! object 'loc' not found
# --> projectdir/input_data/demo/new_haven_table_2025.csv,
#       projectdir/2025_headings.txt
glue_here("{year}_headings_{topic}.txt", topic = "education")
#> Error in purrr::map(x, function(x0) {    glue_here_once_(x0, ..., .snake = .snake, .envir = .GlobalEnv)}): ℹ In index: 1.
#> Caused by error:
#> ! Failed to evaluate glue component {year}
#> Caused by error:
#> ! object 'year' not found
# --> projectdir/2025_headings_education.txt