Median age by sex

This is a sample of data from the 2023 ACS, table B01002. Values are median age for Connecticut, the Capitol Region and Greater Bridgeport councils of government (COGs), and all towns in either COG, each for total population, males, and females.

Usage

median_age

Format

A data frame with 141 rows and 5 variables:

level

Geographic level: factor with levels for state, county, and town

county

County: NA for state- and county- level observations, county name for town-level observations

name

Name of geography

sex

Sex: factor with levels total, male, and female

value

Median age

Source

2023 US Census Bureau American Community Survey 5-year estimates

Examples

head(median_age)
#> # A tibble: 6 × 5
#>   level  county name        sex    value
#>   <fct>  <chr>  <chr>       <fct>  <dbl>
#> 1 state  NA     Connecticut total   41.2
#> 2 state  NA     Connecticut male    39.8
#> 3 state  NA     Connecticut female  42.5
#> 4 county NA     Capitol COG total   40.2
#> 5 county NA     Capitol COG male    38.7
#> 6 county NA     Capitol COG female  41.7