Make multi-hue, multi-shade qualitative color palettes

palx is based on the Palx javascript library, basically a port of its core functions using the colorspace package. Provided a base color, it pulls a set of hues from across the spectrum with the same lightness and saturation, plus a gray color. It then creates shades of those hues from nearly black to nearly white. You probably don't actually want every hue; it's just a good way to get a bunch of colors to choose from.

plot_palx makes a grid with ggplot of hue vs shade tiles. This can also be done with palx, setting plot = TRUE.

Usage

palx(
  hex,
  n_hues = 8,
  n_shades = 9,
  row = NULL,
  as_df = FALSE,
  plot = FALSE,
  labels = FALSE
)

plot_palx(data, labels = FALSE)

Source

https://github.com/jxnblk/palx

Arguments

hex

A string of a color hex code (e.g. "#6f54d6"). This should be a single value, as it's the color the palette will be based around, though it's not guaranteed that this exact color will be in the final palette.

n_hues

A number between 1 and 12. The number of hues returned will be this + 1, because a gray color will be added.

n_shades

A number, for the number of shades to return for each hue. Defaults 9; too many more than that will probably become hard to distinguish.

row

Numeric. If NULL, the default, all rows (shades) are returned. Otherwise, only the rows with these indices are returned. Just a shortcut for subsetting the list.

as_df

Logical, whether to return a data frame (as_df = TRUE) or a list of character vectors. Defaults FALSE.

plot

Logical, whether to call plot_palx before returning. This doesn't change what the function returns, it just prints out a ggplot chart and returns the colors as normal. Defaults FALSE.

labels

Logical, whether to add labels on each tile giving colors' hex codes. Defaults FALSE.

data

The output of calling palx, as either a list or data frame

Value

If as_df = TRUE, a data frame with n_shades rows by one column per hue, plus a column giving the shade number. Otherwise, a named list (length n_shades) of character vectors, where each list item represents one shade.

Details

Some notes about color:

Hue refers to what would commonly be the general name of a color, e.g. blue or yellow.

Shade refers to how light or dark a color is, e.g. light blue, dark blue.

Saturation refers to the strength of a color, or its distance from a neutral gray.

The color spectrum is broken up into even chunks, but there's a set of wavelengths that appear pretty similar--this shows up in the lime-green-teal-cyan-blue colors. That's just how human eyeballs work I guess.

These palettes shouldn't be used for sequential palettes, even pulling multiple shades from one hue. Good sequential palettes are corrected for perceptual differences in shade, whereas this just takes even steps. Also please don't use rainbow scales for sequential data.

You likely only want one or two shades worth of colors (use the row arguments), either adjacent for colors that appear "even", or one light and one dark for a paired palette. Saturated colors that aren't too dark or too light work best.

See also

Examples

palx("#9CCC0C") # returns a list
#> $shade01
#>     green      cyan      blue    violet   fuschia       red    orange      lime 
#> "#022902" "#02291F" "#021629" "#0C0229" "#290229" "#29020C" "#291602" "#1F2902" 
#>      gray 
#> "#171813" 
#> 
#> $shade02
#>     green      cyan      blue    violet   fuschia       red    orange      lime 
#> "#055405" "#055440" "#052D54" "#190554" "#540554" "#540519" "#542D05" "#405405" 
#>      gray 
#> "#2F3227" 
#> 
#> $shade03
#>     green      cyan      blue    violet   fuschia       red    orange      lime 
#> "#088008" "#088062" "#084480" "#260880" "#800880" "#800826" "#804408" "#628008" 
#>      gray 
#> "#484C3B" 
#> 
#> $shade04
#>     green      cyan      blue    violet   fuschia       red    orange      lime 
#> "#0AAB0A" "#0AAB83" "#0A5AAB" "#320AAB" "#AB0AAB" "#AB0A32" "#AB5A0A" "#83AB0A" 
#>      gray 
#> "#606650" 
#> 
#> $shade05
#>     green      cyan      blue    violet   fuschia       red    orange      lime 
#> "#0DDA0D" "#0DDAA7" "#0D73DA" "#400DDA" "#DA0DDA" "#DA0D40" "#DA730D" "#A7DA0D" 
#>      gray 
#> "#7B8265" 
#> 
#> $shade06
#>     green      cyan      blue    violet   fuschia       red    orange      lime 
#> "#32F332" "#32F3C3" "#3293F3" "#6232F3" "#F332F3" "#F33262" "#F39332" "#C3F332" 
#>      gray 
#> "#99A086" 
#> 
#> $shade07
#>     green      cyan      blue    violet   fuschia       red    orange      lime 
#> "#6DF66D" "#6DF6D4" "#6DB2F6" "#906DF6" "#F66DF6" "#F66D90" "#F6B26D" "#D4F66D" 
#>      gray 
#> "#B7BCA9" 
#> 
#> $shade08
#>     green      cyan      blue    violet   fuschia       red    orange      lime 
#> "#A8FAA8" "#A8FAE5" "#A8D1FA" "#BDA8FA" "#FAA8FA" "#FAA8BD" "#FAD1A8" "#E5FAA8" 
#>      gray 
#> "#D4D7CC" 
#> 
#> $shade09
#>     green      cyan      blue    violet   fuschia       red    orange      lime 
#> "#E3FDE3" "#E3FDF7" "#E3F0FD" "#EAE3FD" "#FDE3FD" "#FDE3EA" "#FDF0E3" "#F7FDE3" 
#>      gray 
#> "#F1F2EF" 
#> 
palx("#9CCC0C", as_df = TRUE) # returns a data frame
#> # A tibble: 9 × 10
#>   shade green   cyan    blue    violet  fuschia red     orange  lime    gray   
#>   <dbl> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1     1 #022902 #02291F #021629 #0C0229 #290229 #29020C #291602 #1F2902 #171813
#> 2     2 #055405 #055440 #052D54 #190554 #540554 #540519 #542D05 #405405 #2F3227
#> 3     3 #088008 #088062 #084480 #260880 #800880 #800826 #804408 #628008 #484C3B
#> 4     4 #0AAB0A #0AAB83 #0A5AAB #320AAB #AB0AAB #AB0A32 #AB5A0A #83AB0A #606650
#> 5     5 #0DDA0D #0DDAA7 #0D73DA #400DDA #DA0DDA #DA0D40 #DA730D #A7DA0D #7B8265
#> 6     6 #32F332 #32F3C3 #3293F3 #6232F3 #F332F3 #F33262 #F39332 #C3F332 #99A086
#> 7     7 #6DF66D #6DF6D4 #6DB2F6 #906DF6 #F66DF6 #F66D90 #F6B26D #D4F66D #B7BCA9
#> 8     8 #A8FAA8 #A8FAE5 #A8D1FA #BDA8FA #FAA8FA #FAA8BD #FAD1A8 #E5FAA8 #D4D7CC
#> 9     9 #E3FDE3 #E3FDF7 #E3F0FD #EAE3FD #FDE3FD #FDE3EA #FDF0E3 #F7FDE3 #F1F2EF

# bad examples all with one hue that return weird / not very useful palettes
palx("#ccf4fa") # saturated but too light
#> $shade01
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#081253" "#360853" "#530837" "#531008" "#534908" "#255308" "#085324" "#084B53" 
#>      gray 
#> "#293232" 
#> 
#> $shade02
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#1125AB" "#6E11AB" "#AB1172" "#AB2111" "#AB9611" "#4DAB11" "#11AB4A" "#119AAB" 
#>      gray 
#> "#546868" 
#> 
#> $shade03
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#3149EB" "#A231EB" "#EB31A6" "#EB4531" "#EBD331" "#7AEB31" "#31EB76" "#31D7EB" 
#>      gray 
#> "#829A9A" 
#> 
#> $shade04
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#8997F3" "#CA89F3" "#F389CC" "#F39489" "#F3E689" "#B3F389" "#89F3B0" "#89E8F3" 
#>      gray 
#> "#B7C5C5" 
#> 
#> $shade05
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#CFD4FA" "#E9CFFA" "#FACFEA" "#FAD3CF" "#FAF5CF" "#E0FACF" "#CFFADF" "#CFF6FA" 
#>      gray 
#> "#E2E7E7" 
#> 
#> $shade06
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#D9DEFB" "#EED9FB" "#FBD9EF" "#FBDDD9" "#FBF7D9" "#E7FBD9" "#D9FBE6" "#D9F8FB" 
#>      gray 
#> "#E8EDED" 
#> 
#> $shade07
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#E4E7FC" "#F3E4FC" "#FCE4F3" "#FCE7E4" "#FCF9E4" "#EEFCE4" "#E4FCED" "#E4FAFC" 
#>      gray 
#> "#EFF2F2" 
#> 
#> $shade08
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#EFF1FD" "#F8EFFD" "#FDEFF8" "#FDF1EF" "#FDFCEF" "#F5FDEF" "#EFFDF4" "#EFFCFD" 
#>      gray 
#> "#F5F7F7" 
#> 
#> $shade09
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#FAFBFE" "#FDFAFE" "#FEFAFD" "#FEFAFA" "#FEFEFA" "#FCFEFA" "#FAFEFC" "#FAFEFE" 
#>      gray 
#> "#FCFDFD" 
#> 
palx("#6c888d") # not too light, but not saturated enough
#> $shade01
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#16161C" "#1A161C" "#1C161A" "#1C1716" "#1C1C16" "#181C16" "#161C18" "#161B1C" 
#>      gray 
#> "#161B1C" 
#> 
#> $shade02
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#2D2E3A" "#352D3A" "#3A2D35" "#3A2F2D" "#3A392D" "#323A2D" "#2D3A32" "#2D383A" 
#>      gray 
#> "#2D383A" 
#> 
#> $shade03
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#444558" "#514458" "#584450" "#584744" "#585644" "#4B5844" "#44584C" "#445558" 
#>      gray 
#> "#445458" 
#> 
#> $shade04
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#5A5D76" "#6C5A76" "#765A6B" "#765F5A" "#76745A" "#65765A" "#5A7665" "#5A7276" 
#>      gray 
#> "#5B7175" 
#> 
#> $shade05
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#737693" "#887393" "#937387" "#937873" "#939073" "#7F9373" "#739380" "#738E93" 
#>      gray 
#> "#748D92" 
#> 
#> $shade06
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#9294AC" "#A292AC" "#AC92A1" "#AC9692" "#ACA992" "#9BAC92" "#92AC9C" "#92A8AC" 
#>      gray 
#> "#93A7AB" 
#> 
#> $shade07
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#B1B3C4" "#BDB1C4" "#C4B1BC" "#C4B4B1" "#C4C2B1" "#B8C4B1" "#B1C4B9" "#B1C1C4" 
#>      gray 
#> "#B2C0C3" 
#> 
#> $shade08
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#D1D2DC" "#D8D1DC" "#DCD1D7" "#DCD2D1" "#DCDBD1" "#D5DCD1" "#D1DCD5" "#D1DADC" 
#>      gray 
#> "#D1DADB" 
#> 
#> $shade09
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#F0F1F4" "#F2F0F4" "#F4F0F2" "#F4F1F0" "#F4F3F0" "#F1F4F0" "#F0F4F2" "#F0F3F4" 
#>      gray 
#> "#F0F3F4" 
#> 
palx("#0e91a7") # much better--high saturation, middle lightness
#> $shade01
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#030621" "#160321" "#210316" "#210703" "#211E03" "#0E2103" "#03210F" "#031D21" 
#>      gray 
#> "#101414" 
#> 
#> $shade02
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#060D45" "#2E0645" "#45062D" "#450E06" "#453D06" "#1D4506" "#06451E" "#063D45" 
#>      gray 
#> "#21292A" 
#> 
#> $shade03
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#091468" "#450968" "#680944" "#681509" "#685D09" "#2C6809" "#09682E" "#095C68" 
#>      gray 
#> "#313E40" 
#> 
#> $shade04
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#0C1B8C" "#5D0C8C" "#8C0C5A" "#8C1C0C" "#8C7D0C" "#3B8C0C" "#0C8C3D" "#0C7B8C" 
#>      gray 
#> "#425355" 
#> 
#> $shade05
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#0F23B6" "#790FB6" "#B60F76" "#B6250F" "#B6A30F" "#4CB60F" "#0FB650" "#0FA0B6" 
#>      gray 
#> "#566C6F" 
#> 
#> $shade06
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#1F37EC" "#A11FEC" "#EC1F9D" "#EC3A1F" "#ECD41F" "#6AEC1F" "#1FEC6E" "#1FD1EC" 
#>      gray 
#> "#769195" 
#> 
#> $shade07
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#6071F2" "#BC60F2" "#F260B9" "#F27360" "#F2E060" "#95F260" "#60F298" "#60DFF2" 
#>      gray 
#> "#9EB1B4" 
#> 
#> $shade08
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#A0AAF7" "#D7A0F7" "#F7A0D6" "#F7ACA0" "#F7EDA0" "#C0F7A0" "#A0F7C2" "#A0ECF7" 
#>      gray 
#> "#C5D0D2" 
#> 
#> $shade09
#>    indigo   fuschia      pink    orange    yellow     green      teal      blue 
#> "#E1E4FC" "#F2E1FC" "#FCE1F2" "#FCE4E1" "#FCF9E1" "#EBFCE1" "#E1FCEB" "#E1F9FC" 
#>      gray 
#> "#ECF0F1" 
#>